move net liquidity tooltip to apy label
