+++ src/hooks/useUserHealth.ts
@@ -12,7 +12,7 @@ export function useUserHealth() {
   const dispatch = useAppDispatch();
   const { showDailyReturns } = useAppSelector(getAppState);
   const netAPY = useAppSelector(getNetAPY({ isStaking: false }));
-  const netTvlAPY = useAppSelector(getNetTvlAPY({ isStaking: false }));
+  const netLiquidityAPY = useAppSelector(getNetTvlAPY({ isStaking: false }));
   const dailyReturns = useAppSelector(getDailyReturns);
   const healthFactor = useAppSelector(getHealthFactor);
   const { fullDigits, setDigits } = useFullDigits();
@@ -29,7 +29,7 @@ export function useUserHealth() {
 
   return {
     netAPY,
-    netTvlAPY,
+    netLiquidityAPY,
     dailyReturns,
     healthFactor,
     slimStats,
